rename: 훅 폴더 이름 직관적으로 변경
